# Update to webpack@4