--- a/browser/webpack.debug.config.js
+++ b/browser/webpack.debug.config.js
@@ -6,15 +6,18 @@ const webpack = require("webpack")
 const baseConfig = require("./webpack.development.config.js")
 
 const debugConfig = Object.assign({}, baseConfig, {
+    devtool: "source-map",
     plugins: [
         new webpack.DefinePlugin({
             "process.env.NODE_ENV": '"development"',
         }),
-        new webpack.optimize.CommonsChunkPlugin({
-            async: true,
-            minChunks: 2,
-        }),
     ],
+    optimization: {
+        splitChunks: {
+            name: "vendor",
+            minChunks: 2,
+        },
+    },
     output: {
         path: path.join(__dirname, "..", "lib", "browser"),
         publicPath: "lib/browser/",