## Before

- `npm run build`: 64239ms
- `npm start`: 38832ms
- `du -h lib/browser`: 4.9M

## After

- `npm run build`: 27127ms
- `npm start`: 35269ms
- `du -h lib/browser`: 4.9M

Note: babili is not needed anymore since webpack4 has a minifier by default (uglify-es) and optimize-js optimization is old, since browsers have fixed this bug (and the plugin optimize-js-plugin has not been update to webpack 4)