@Akin909 2018-04-17
@DeltaEvo Hmm seems to be looking for the `sanitize` method on `marked` which we use to render markdown but that is in no way related to the prettier plugin, we dont use or require marked anywhere near the prettier one, but the prettier plugin is built right after the `oni-markdown-plugin` so not sure if the error is coming from elsewhere but appearing in that test.

If you run the plugin locally does it work correctly or does it throw the same/ a similar error?