@DeltaEvo 2018-04-17
@Akin909 locally the plugin seems to work, i will try to look into `oni-markdown-plugin` (I was thinking that marked was used by prettier :joy: )