@DeltaEvo 2018-04-18
Well, tests saved us from a very strange bug, thanks @bryphe and @Akin909 (and all other) for these tests :heart: