@bryphe 2018-04-18
Nice catch, @DeltaEvo ! Glad you were able to figure out the issue 👍 And ya, definitely glad we had test coverage around that - nice work on the CiTest @Akin909 !