@bryphe 2018-04-18
Just looked at the snippets test issue - looks like there is a crash here:
```
        const detectIndentPromise = import("detect-indent")

        const [bufferLines, detectIndent] = await Promise.all([
            bufferLinesPromise,
            detectIndentPromise,
        ])

        const ret = detectIndent(bufferLines.join("\n"))
```

The `detectIndent` method can't be found. I can repro this just by adding a snippet (don't need to run the test, just try expanding the `import snippet` in a typescript/javascript file)