@bryphe 2018-04-18
Looks like the new `import` behavior is working a bit differently:
![image](https://user-images.githubusercontent.com/13532591/38907069-da476344-426e-11e8-96d6-81f2ee8c2f97.png)

Before, `detectIndent` was a function, but now it has a `default` member - I wonder if there is a difference in how it resolves/splits async imports.

A quick fix that resolved this locally was to just do:
```
        const ret = (detectIndent as any).default(bufferLines.join("\n"))
```
Unfortunately, it requires a cast to `any` - the typings for `detect-indent` assume the function is exported, which is out of sync with what the `import` is doing. Not sure if this is a webpack bug, or something we are doing wrong. I'm okay with the `any` fix though, just might need a comment to explain why its needed 😄