@DeltaEvo 2018-04-18
Thanks for investigating at this @bryphe this new behavior is described here https://medium.com/webpack/webpack-4-import-and-commonjs-d619d626b655
To avoid using any and so loose typing, I think i will remove the dynamic import on `detect-indent`, it's a fairly small module so using a dynamic import on it seems to add more overhead that something else :)