@Akin909 2018-04-18
@DeltaEvo I agree re. `detect-indent` 👍, this branch looks good now and all tests are passing so will bring this in if you're done, and quick check with @bryphe if you have any objections?